Implemented: PartyName from util module to party module(#121)
