+++ src/services/PartyService.ts
@@ -1,0 +1,13 @@
+import { api } from '@/adapter';
+
+const fetchPartyInformation = async (query: any): Promise <any>  => {
+    return api({
+      url: "performFind",
+      method: "get",
+      params: query
+    });
+  }
+
+  export const partyService = {
+    fetchPartyInformation
+  }
