+++ src/services/UtilService.ts
@@ -280,14 +280,6 @@ const fetchCarrierPartyIds = async (query: any): Promise <any>  => {
   });
 }
 
-const fetchPartyInformation = async (query: any): Promise <any>  => {
-  return api({
-    url: "performFind",
-    method: "get",
-    params: query
-  });
-}
-
 const fetchShipmentMethodTypeDesc = async (query: any): Promise <any>  => {
   return api({
     url: "performFind",
@@ -349,7 +341,6 @@ export const UtilService = {
   fetchCarrierPartyIds,
   fetchDefaultShipmentBox,
   fetchFacilityTypeInformation,
-  fetchPartyInformation,
   fetchPicklistInformation,
   fetchRejectReasons,
   fetchShipmentBoxTypeDesc,
