+++ src/store/modules/party/mutation-type.ts
@@ -1,0 +1,2 @@
+export const SN_PARTY = 'party'
+export const PARTY_PARTY_NAMES_UPDATED = SN_PARTY + '/PARTY_NAMES_UPDATED'
