+++ src/store/modules/party/mutations.ts
@@ -1,0 +1,10 @@
+import { MutationTree } from 'vuex'
+import * as types from './mutation-type'
+import partyState from './partyState'
+
+const mutations: MutationTree <partyState> = {
+    [types.PARTY_PARTY_NAMES_UPDATED] (state, payload) {
+        state.partyNames = payload
+      },
+}
+export default mutations;
