+++ src/store/modules/party/partyState.ts
@@ -1,0 +1,3 @@
+export default interface partyState {
+    partyNames: any;
+}
