+++ src/store/modules/util/UtilState.ts
@@ -1,6 +1,5 @@
 export default interface UtilState {
   rejectReasons: [];
-  partyNames: any;
   shipmentMethodTypeDesc: any;
   shipmentBoxTypeDesc: any;
   facilityTypeDesc: any;
