+++ src/store/modules/util/index.ts
@@ -9,7 +9,6 @@ const utilModule: Module<UtilState, RootState> = {
   namespaced: true,
   state: {
     rejectReasons: [],
-    partyNames: {},
     shipmentMethodTypeDesc: {},
     shipmentBoxTypeDesc: {},
     facilityTypeDesc: {},
